+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-04-25 22:22:07
-// Commit: chore(i18n): update dependencies and cleanup unused imports in i18n
+// Updated at: 2026-04-26 17:32:34
+// Commit: test(i18n): add unit tests for i18n business logic
 
 class i18nComponentState {
-  final String commitRef = '693';
-  final String lastMessage = 'chore(i18n): update dependencies and cleanup unused imports in i18n';
-  final DateTime timestamp = DateTime.parse('2026-04-25T22:22:07+05:00');
+  final String commitRef = '717';
+  final String lastMessage = 'test(i18n): add unit tests for i18n business logic';
+  final DateTime timestamp = DateTime.parse('2026-04-26T17:32:34+05:00');
 }