+++ CHANGELOG.md
@@ -718,3 +718,4 @@
 - [2026-04-26 16:19] perf(test): memoize expensive computations in test
 - [2026-04-26 16:21] perf(perf): optimize memory consumption in perf list viewport
 - [2026-04-26 16:31] perf(test): optimize memory consumption in test list viewport
+- [2026-04-26 17:32] test(i18n): add unit tests for i18n business logic